pyproject.toml
[project]
name = "movie-rating"
version = "0.6.0"
description = "A REST API for rating movies, built with FastAPI and async SQLAlchemy"
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "psycopg[binary]>=3.3.3",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "sqlalchemy[asyncio]>=2.0.49",
]

[dependency-groups]
dev = [
    "alembic>=1.18.4",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.6",
    "mypy>=1.20.1",
    "pre-commit>=4.5.1",
    "pymdown-extensions>=10.21.2",
    "ruff>=0.15.10",
    "taskipy>=1.14.1",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "aiosqlite>=0.22.1",
    "commitizen>=4.13.10",
]

[tool.ruff]
line-length = 79
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".gitignore",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "__pycache__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "alembic",
    "migrations",
]

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['PLR2004', 'PLR0917', 'PLR0913']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.mypy]
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.commitizen]
version_provider = "uv"
tag_format = "v$version"
update_changelog_on_bump = true

[tool.taskipy.tasks]
lint = 'ruff check'
lint_fix = 'ruff check --fix'
format = 'ruff format'
post_format = 'task lint_fix'
mypy = 'mypy src/'
test_check = 'pytest -s -x -vv'
test = 'pytest -s -x --cov=src/ -vv'
post_test = 'coverage html'
pre_app = 'task lint_fix && task format && task mypy && task test_check && docker compose up -d'
app = 'fastapi dev app.py'
